## Summary
- Adds `make dev-exp` target that reduces compilation time from 2m 24s to 1m 23s (42% improvement)
- Maintains full functionality with all tests passing
- Provides faster development iteration cycles

## Changes
- **New Makefile target**: `make dev-exp` for experimental fast builds
- **Optimization approach**: Uses `-O0` instead of `-O3` for development builds
- **Build system updates**: 
  - Modified `Makefile.gfortran` to support fast profile
  - Simplified `meson.build` flags for plain builds
  - Added parallel compilation support
- **Documentation**: Added `tools/` directory with optimization guide and benchmarks

## Test Results
✅ All tests pass with the optimized build:
- 20 tests passed, 4 skipped
- Core functionality verified (water balance, energy fluxes, multi-grid simulations)
- No regression in simulation accuracy

## Usage
```bash
# Standard build (slower)
make dev          # 2m 24s

# Optimized build (faster) 
make dev-exp      # 1m 23s (42% faster)

# Ultra-fast build (experimental)
make dev-ultrafast
```

## Benchmark Details
See `tools/OPTIMIZATION_SUMMARY.md` for detailed benchmark results and optimization techniques.